refactor test modules for scraper: improve docstrings and add shared helper tests
